Add overlays for bamboo logs
Adds overlays to Bamboo Blocks and Stripped Bamboo Blocks, based on the block's orientation. This works with both planks and normal blocks. 

To not add bloat to the files in the planks folders, the bamboo takes priority and overlays itself onto all plank types.
To simplify bamboo to stripped bamboo, stripped bamboo takes priority and overlays itself onto bamboo blocks.